+++ container-chain-pallets/authorities-noting/src/benchmarks.rs
@@ -105,3 +105,26 @@ benchmarks! {
         crate::mock::Test
     );
 }
+
+
+/// Benchmark Helper
+pub trait BenchmarkHelper<AuthorityId> {
+	/// Returns the authorities on empty
+	fn authorities_on_empty() -> Vec<AuthorityId>;
+}
+
+impl<AuthorityId> BenchmarkHelper<AuthorityId> for () {
+    fn authorities_on_empty() -> Vec<AuthorityId> {
+        return vec![]
+    }
+}
+
+pub struct NimbusIdBenchmarkHelper;
+
+use nimbus_primitives::NimbusId;
+use sp_core::ByteArray;
+impl<AuthorityId: From<NimbusId>> BenchmarkHelper<AuthorityId> for NimbusIdBenchmarkHelper {
+    fn authorities_on_empty() -> Vec<AuthorityId> {
+        return vec![NimbusId::from_slice(&[1; 32]).unwrap().into()]
+    }
+}
